project base
here is the current minigauge cam. it includes the first project fork,
where I am trying to build the system on a bare arduino (boarduino) with
more appropriate use of sleep functions. So far, not much good has come
of it. However, the base miniGaugeCam functions great. It would be nice
to add a wireless transfer component...